refactor: 컨트롤러 로직 도메인별로 묶기
